Fix/#490 알림 삭제 버그 수정 (#495)
* refactor(PrimaryNotificationFragment): 한 개의 리사이클러 뷰를 두 개로 나눔

- 최신 알림과 이전 알림을 동시에 가지고 있던 리사이클러 뷰를 두 개의 리사이클러 뷰에서 각각 보여주도록 나눔

* refactor: 다크 모드 테마 제거

* chore(build.gradle.kts): 리소스 축소 해제

* refactor(primaryNotification): 주요 알림 화면 리팩터링

- 알림 목록을 보여주는 기능은 아직 구현하지 않음

* feat(PrimaryNotification): 주요 알림 목록 보여주기 기능 구현

* feat(PrimaryNotification): 주요 알림 읽기 및 삭제 및 화면 이동 기능 구현

* feat(PrimaryNotification): UiEvent 처리 구현

* feat(PrimaryNotification): 주요 알림들이 같이 스크롤 되도록 변경

* refactor(PrimaryNotification): 불필요한 파일 제거 및 패키지화

* chore(build.gradle.kts): 리소스 축소하도록 변경

* style: ktlint 적용

* refactor(PrimaryNotification): Event 적용

* refactor(PrimaryNotificationViewModel): uiState에 설정할 값을 데이터 모델에서 추출하는 로직을 viewModel에서 UiState로 옮김

* refactor: ViewHolder를 재사용하도록 변경

---------

+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/notificationBox/primaryNotification/PrimaryNotificationViewModel.kt
@@ -1,37 +1,33 @@
 package com.emmsale.presentation.ui.notificationBox.primaryNotification
 
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.emmsale.data.common.ApiError
 import com.emmsale.data.common.ApiException
 import com.emmsale.data.common.ApiSuccess
 import com.emmsale.data.notification.NotificationRepository
-import com.emmsale.data.notification.updated.UpdatedNotification
 import com.emmsale.data.token.TokenRepository
 import com.emmsale.presentation.KerdyApplication
+import com.emmsale.presentation.common.Event
 import com.emmsale.presentation.common.livedata.NotNullLiveData
 import com.emmsale.presentation.common.livedata.NotNullMutableLiveData
 import com.emmsale.presentation.common.viewModel.RefreshableViewModel
 import com.emmsale.presentation.common.viewModel.ViewModelFactory
-import com.emmsale.presentation.ui.notificationBox.primaryNotification.uistate.PrimaryNotificationUiState
-import com.emmsale.presentation.ui.notificationBox.primaryNotification.uistate.PrimaryNotificationsUiEvent
-import com.emmsale.presentation.ui.notificationBox.primaryNotification.uistate.PrimaryNotificationsUiState
+import com.emmsale.presentation.ui.notificationBox.primaryNotification.uievent.PrimaryNotificationsUiEvent
+import com.emmsale.presentation.ui.notificationBox.primaryNotification.uistate.PrimaryNotificationScreenUiState
 import kotlinx.coroutines.launch
 
 class PrimaryNotificationViewModel(
     private val tokenRepository: TokenRepository,
     private val notificationRepository: NotificationRepository,
 ) : ViewModel(), RefreshableViewModel {
-    private val _pastNotifications = NotNullMutableLiveData(PrimaryNotificationsUiState())
-    val pastNotifications: NotNullLiveData<PrimaryNotificationsUiState> = _pastNotifications
 
-    private val _recentNotifications = NotNullMutableLiveData(PrimaryNotificationsUiState())
-    val recentNotifications: NotNullLiveData<PrimaryNotificationsUiState> = _recentNotifications
+    private val _uiState =
+        NotNullMutableLiveData<PrimaryNotificationScreenUiState>(PrimaryNotificationScreenUiState.Loading)
+    val uiState: NotNullLiveData<PrimaryNotificationScreenUiState> = _uiState
 
-    private val _event = MutableLiveData<PrimaryNotificationsUiEvent?>(null)
-    val event: LiveData<PrimaryNotificationsUiEvent?> = _event
+    private val _uiEvent = NotNullMutableLiveData(Event(PrimaryNotificationsUiEvent.NONE))
+    val uiEvent: NotNullLiveData<Event<PrimaryNotificationsUiEvent>> = _uiEvent
 
     init {
         refresh()
@@ -40,70 +36,50 @@ class PrimaryNotificationViewModel(
     override fun refresh() {
         viewModelScope.launch {
             val uid = tokenRepository.getToken()?.uid ?: return@launch
-            _recentNotifications.value = recentNotifications.value.copy(isLoading = true)
-
             when (val result = notificationRepository.getUpdatedNotifications(uid)) {
-                is ApiSuccess -> {
-                    updatePastNotifications(result.data.filter { it.isRead })
-                    updateRecentNotifications(result.data.filterNot { it.isRead })
-                }
+                is ApiError, is ApiException ->
+                    _uiState.value =
+                        PrimaryNotificationScreenUiState.Error
 
-                is ApiError, is ApiException -> {
-                    _recentNotifications.value =
-                        recentNotifications.value.copy(isLoading = false, isError = true)
-                }
+                is ApiSuccess ->
+                    _uiState.value =
+                        PrimaryNotificationScreenUiState.Success.from(result.data)
             }
         }
     }
 
-    private fun updateRecentNotifications(notifications: List<UpdatedNotification>) {
-        _recentNotifications.value = PrimaryNotificationsUiState(
-            notifications = notifications.map(PrimaryNotificationUiState::from),
-        )
-    }
-
-    private fun updatePastNotifications(notifications: List<UpdatedNotification>) {
-        _pastNotifications.value = PrimaryNotificationsUiState(
-            notifications = notifications.map(PrimaryNotificationUiState::from),
-        )
-    }
-
-    fun changeToRead(notificationId: Long) {
+    fun deleteAllPastNotifications() {
         viewModelScope.launch {
-            notificationRepository.updateUpdatedNotificationReadStatus(notificationId)
+            val currentUiState = _uiState.value
+            if (currentUiState !is PrimaryNotificationScreenUiState.Success) return@launch
+            val pastNotificationIds = currentUiState.pastNotifications.map { it.notificationId }
+            when (notificationRepository.deleteUpdatedNotifications(pastNotificationIds)) {
+                is ApiSuccess -> refresh()
+                is ApiError, is ApiException ->
+                    _uiEvent.value =
+                        Event(PrimaryNotificationsUiEvent.DELETE_FAIL)
+            }
+        }
+    }
+
+    fun readNotification(notificationId: Long) {
+        viewModelScope.launch {
+            when (notificationRepository.updateUpdatedNotificationReadStatus(notificationId)) {
+                is ApiSuccess -> refresh()
+                is ApiError, is ApiException -> Unit
+            }
         }
     }
 
     fun deleteNotification(notificationId: Long) {
         viewModelScope.launch {
             when (notificationRepository.deleteUpdatedNotifications(listOf(notificationId))) {
-                is ApiSuccess -> deleteNotificationById(notificationId)
-                is ApiError, is ApiException -> updateToNotificationDeleteErrorState()
+                is ApiSuccess -> refresh()
+                is ApiError, is ApiException ->
+                    _uiEvent.value =
+                        Event(PrimaryNotificationsUiEvent.DELETE_FAIL)
             }
         }
-    }
-
-    private fun deleteNotificationById(id: Long) {
-        _pastNotifications.value = pastNotifications.value.deleteNotificationById(id)
-        _recentNotifications.value = recentNotifications.value.deleteNotificationById(id)
-    }
-
-    private fun updateToNotificationDeleteErrorState() {
-        _event.value = PrimaryNotificationsUiEvent.DELETE_ERROR
-    }
-
-    fun deleteAllPastNotifications() {
-        val pastNotificationIds = pastNotifications.value.notificationIds
-        viewModelScope.launch {
-            when (notificationRepository.deleteUpdatedNotifications(pastNotificationIds)) {
-                is ApiSuccess -> _pastNotifications.value = PrimaryNotificationsUiState.EMPTY
-                is ApiError, is ApiException -> updateToNotificationDeleteErrorState()
-            }
-        }
-    }
-
-    fun resetEvent() {
-        _event.value = null
     }
 
     companion object {
